+++ react-components/src/components/CompareMarkets/DataTable.jsx
@@ -1,5 +1,7 @@
-import React, { useState, useEffect } from 'react'
+import React, { useEffect } from 'react'
 import PropTypes from 'prop-types'
+import Services from '@src/Services'
+import actions from '@src/actions'
 import { Tooltip } from '@components/tooltip/Tooltip'
 import { isArray, mapArray, deepAssign } from '../../Helpers'
 
@@ -13,14 +15,14 @@ export default function DataTable(props) {
     comparisonMarkets,
     commodityCode,
     removeMarket,
+    cacheVersion,
   } = props
-  const [lCache, setLCache] = useState(
-    (cache[datasetName] = cache[datasetName] || {})
-  )
 
   const dataIn = (data) => {
     cache[datasetName] = deepAssign(cache[datasetName], data)
-    setLCache(cache[datasetName])
+    Services.store.dispatch(
+      actions.setLoaded()
+    )
   }
 
   const flagArray = (array, value) => {
@@ -36,7 +38,9 @@ export default function DataTable(props) {
     // Set the columns in the countries provided - to loading
     const loadingIndicators = {}
     countries.forEach((country) => {
-      loadingIndicators[country] = { loading: flagArray(columnList, 1) }
+      loadingIndicators[country.country_iso2_code] = {
+        loading: flagArray(columnList, 1),
+      }
     })
     dataIn(loadingIndicators)
   }
@@ -47,21 +51,31 @@ export default function DataTable(props) {
     return new Promise((resolve, reject) => {
       requestFunction(countries, cache.commodityCode)
         .then((result) => {
-          let data = result
-          if (isArray(data)) {
-            data = mapArray(data, 'country')
+          const outData = {}
+          let inData = result
+          if (isArray(inData)) {
+            inData = mapArray(inData, 'country')
           }
           countries.forEach((country) => {
-            data[country] = data[country] || {}
-            data[country].loading = flagArray(columnList, 0)
+            const iso2 = country.country_iso2_code
+            outData[iso2] = outData[iso2] || 
+                inData[iso2] ||
+                inData[country.country_name] ||
+                {}
+            outData[iso2].loading = flagArray(
+              columnList,
+              0
+            )
           })
-          dataIn(data)
+          dataIn(outData)
           resolve()
         })
         .catch(() => {
           const clear = {}
           countries.forEach((country) => {
-            clear[country] = { loading: flagArray(columnList, 0) }
+            clear[country.country_iso2_code] = {
+              loading: flagArray(columnList, 0),
+            }
           })
           dataIn(clear)
           reject()
@@ -116,17 +130,18 @@ export default function DataTable(props) {
     // Wipe cache if commodity code changes
     if (cache.commodityCode !== commodityCode) {
       cache = { commodityCode }
+      cache[datasetName] = {}
     }
     cache[datasetName] = cache[datasetName] || {}
-    const countries = Object.values(comparisonMarkets).map(
-      (country) => country.country_name
+
+    const missingCountries = Object.values(comparisonMarkets).filter(
+      (country) => !cache[datasetName][country.country_iso2_code]
     )
-    const missingCountries = countries.filter((country) => !lCache[country])
-    setLCache(lCache)
 
     if (missingCountries.length) {
       getTableData(missingCountries)
     }
+
   }, [commodityCode, comparisonMarkets])
 
   const yearDiv = (year, baseYear) => {
@@ -170,7 +185,7 @@ export default function DataTable(props) {
   const years = {}
   Object.values(comparisonMarkets).forEach((market) => {
     const countryData =
-      cache[datasetName] && cache[datasetName][market.country_name]
+      cache[datasetName] && cache[datasetName][market.country_iso2_code]
     if (countryData) {
       Object.values(config.columns).forEach((columnConfig) => {
         if (columnConfig.year) {
@@ -205,7 +220,7 @@ export default function DataTable(props) {
 
   const tableBody = Object.values(comparisonMarkets).map((market) => {
     const countryData =
-      cache[datasetName] && cache[datasetName][market.country_name]
+      cache[datasetName] && cache[datasetName][market.country_iso2_code]
     const countryRow = Object.keys(config.columns).map((columnKey) => {
       const cellConfig = config.columns[columnKey]
       return (
@@ -249,7 +264,7 @@ export default function DataTable(props) {
 
   return (
     <span>
-      <table className="m-v-0 border-blue-deep-20 valign-middle">
+      <table className={`m-v-0 border-blue-deep-20 valign-middle cache-version-${cacheVersion}`}>
         <thead>
           <tr>
             <th className="body-l-b">&nbsp;</th>
@@ -302,4 +317,8 @@ DataTable.propTypes = {
   }).isRequired,
   commodityCode: PropTypes.string.isRequired,
   removeMarket: PropTypes.func.isRequired,
+  cacheVersion: PropTypes.number,
+}
+DataTable.defaultProps = {
+  cacheVersion: null
 }